api/requirements.txt
# Flask API Requirements
flask==3.1.2
flask-cors==5.0.0
pyjwt==2.10.1
requests==2.32.5
python-dotenv==1.2.1
motor==3.7.1
pymongo==4.15.3
waitress==3.0.2
